requirements.txt
a2wsgi==1.10.8
aiodns==3.4.0
aiohttp_socks==0.10.1
async_timeout==5.0.1
Babel==2.17.0
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.2
curio==1.6
cython==3.1.1
docutils==0.21.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi_cli==0.0.7
filelock==3.18.0
gunicorn==23.0.0
h2==4.2.0
HTMLParser==0.0.2
httptools==0.6.4
hypothesis==6.131.21
ipython==9.2.0
ipywidgets==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.6
jnius==1.1.0
js==1.0
keyring==25.6.0
mask==1.0.0
motor==3.7.1
ollama==0.4.8
orjson==3.10.18
outcome==1.3.0.post0
protobuf==6.31.0
pyodide==0.0.2
pyOpenSSL==25.1.0
pytest==8.3.5
python_multipart==0.0.20
PyYAML==6.0.2
qdrant_client==1.14.2
sentence_transformers==4.1.0
simplejson==3.20.1
socksio==1.0.0
Sphinx==8.2.3
thread==2.0.5
toml==0.10.2
torch==2.7.0
transformers==4.52.2
ujson==5.10.0
urllib3_secure_extra==0.1.0
watchfiles==1.0.5
wsproto==1.2.0
xmlrpclib==1.0.1
zstandard==0.23.0
